requirements.txt
flask
flask-sqlalchemy
python-dotenv
google-generativeai==0.8.6
Pillow
reportlab
